requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
autoenv==1.0.0
cachetools==3.1.0
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
discord.py==1.0.1
google-api-core==1.10.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==0.29.1
google-cloud-error-reporting==0.30.1
google-cloud-logging==1.10.0
google-cloud-profiler==1.0.8
google-python-cloud-debugger==2.10
googleapis-common-protos==1.5.10
grpcio==1.20.1
httplib2==0.12.3
idna==2.8
multidict==4.5.2
protobuf==3.7.1
psycopg2==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyNaCl==1.2.1
pytz==2019.1
PyYAML==5.1
requests==2.21.0
rollbar==0.14.7
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.3
uritemplate==3.0.0
urllib3==1.24.3
websockets==6.0
yarl==1.3.0
elasticsearch-dsl>=6.0.0,<7.0.0
